Don't ignore values that change and then change back
This commit should properly account for situations where we are using
the hash field to track changes to a couple of fields that may change
and then change back to a previous combination (generating the same
hash value). The current code only checks whether the present
combination appears _anywhere_ in that record's ${name}_updates table,
not limiting to the most recent values for that uniqueKey.

E.g. take the following three states

```
uid   name        language    country   updated_date
663   Michael     en          USA       2001-01-01
663   michael s   es          USA       2002-02-02
663   Michael     en          USA       2003-03-03
```

This is a realistic change pattern for real-world data, and if we are
tracking the hash of (name, language, country), we will record the
first state, the second state, and then we will ignore the third.

This commit uses `any_value(${hash} having max ${timstamp}) ... group by ${uniqueKey}`
to make sure we are only looking at the most recent record in the history
table before we compare hashes.