Switch make the fix work on all redshift and snowflakedb too
Since any_value is a bigquery-specific function, this commit switches
to using first_value with a window, partitioning by the uniqueKey, and
ordering by the timestamp.

I have tested this approach on bigquery, and from the docs it appears
the first_value and window and frame definitions work the same across
all three platforms, but I haven't tested with live data on the other
two.